pyproject.toml
[tool.poetry]
name = "core"
version = "0.1.0"
description = ""
authors = ["Queenie Grace Isidro <queeniegraceisidro@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
django = "^5.0"
djangorestframework = "^3.14.0"
django-cors-headers = "^4.3.1"
python-decouple = "^3.8"
dj-rest-auth = {extras = ["with-social"], version = "^6.0.0"}
djangorestframework-simplejwt = "^5.3.1"
psycopg2-binary = "^2.9.9"
channels = {extras = ["daphne"], version = "^4.1.0"}
channels-redis = "^4.2.0"

[tool.poetry.group.test.dependencies]
pytest-django = "^4.7.0"
factory-boy = "^3.3.0"

[tool.poetry.group.dev.dependencies]
django-naomi = "^0.8"

[tool.pytest.ini_options]
DJANGO_SETTINGS_MODULE = "core.settings"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
